backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
scikit-learn
sqlalchemy
pydantic
python-multipart
python-jose
passlib[bcrypt]
python-dotenv
google-generativeai